Fix the int vs size_t warnings once and for all.
Also add -std=c++11 flag.

+++ class/FileMgr.h
@@ -43,7 +43,7 @@ public:
 	{ return(++_nextCnt == _numberFiles ? (DataFile *)NULL : _dataFile[_nextCnt]); }
 
 private:
-  long int	_numberFiles;
+  size_t	_numberFiles;
   size_t	_currentFile;
   size_t	_nextCnt;
